Implement write-class and wire it into eval, load
This commit implements a write-class function (extracted from Oxcart)
and wires it into the eval and load implementations as appropriate
creating :compile-files kw options on both functions the default value
for which is false.

Fixes TEMJVM-15

Default to false iff *compile-files* isn't bound

This changes the behavior of the :compile-files keyword option so that
rather than defaulting to false if unset, if *compile-files* is bound
that value will be propigated. Only if :compile-files is not provided
and *compile-files* is unbound will the default value of false occur.